Adjust default value for $stomp_pool.
Prior to this commit, the parameter stomp_pool was set to an empty
hash. Later in the class, there is conditional logic validating
user input to that paramater, matching on stomp_pool='UNSET'.

This commit modifies the default value to stomp_pool from {} to
UNSET in order to accept the default behavior of a single pool.