Refactor how plugins are handled.
Prior to this commit, the server and client classes had been
refactored to compress distributed classes into single server and
client classes. This left some cruft that prevented the following
declaration from functioning properly.

class { 'mcollective':
  manage_plugins => true,
}

This commit refactors the class naming and relationships between
resources in order for the above declaration to function reasonably.
It has not been tested and should not be merged without its sibling
refactoring commits.