Simplify mcollective class
Previously, the mcollective class accepted a large number of parameters
that would be evaluated and used to configure all subcomponents of
MCollective.  This made the mcollective class unwieldy, but also led to
some confusion once enterprise support was added. To ease the process
in getting MCollective setup, it was decided to expose all functional
parameters to the main mcollective class (i.e. Should MCollective be
setup with Puppet Enterprise? Should the MCollective client component
be installed? Should this class manage the installation of MCollective
plugins? etc...) and let the mcollective::params be the entry point for
more granular configuration changes.

This commit updates the main mcollective class so that it only takes five
functional parameters: enterprise, client, server, version, and
manage_plugins.