Pass the manage_plugins param to the plugins class
The mcollective::plugins class is currently being refactored to lay
down the default folder structure for all MCollective plugins, but
NOT manage/install the plugins unless it is told to do so. Because
of this change, the manage_plugins parameter needs to be passed to
the mcollective::plugins class so it can perform this conditional
check.  This commit enables this functionality.