Convert params.pp to paramertized class.
Prior to this commit, the mcollective::params class set a bunch of
parameter defaults but did not do so as a parameterized class. This
limits the user interface for overriding these defaults to accomodate
for different scnearios.

This commit converts the class into a parameterized class and alters
some values in the client template to use the fully qualified variable
lookup.