Refactor parameters and templates
Prior to this commit, templates did not have their variable references
fully scoped. Also, prior to this commit, some of the variable defaults
and logic were not aligned for ease of use by the user.

This commit adjusts the parameter logic, sets sane default values and
adjusts the client and server templates to use fully qualified variables
via scope.lookupvar().