Validate stomp_pool correctly
There is a bug where variables in a class's parameter declaration
that refer to other variables in a class's parameter declaration
are subject to parse-order.  Because the order that variables are
evaluated in the parameter declaration section of a class declaration
cannot be reliably determined, it's better to default the value to
'UNSET' and check for that value in the body of the class (conditional
statements in the body of the class can be certain that all variables
in the parameter declaration section of a class declaration have been
evaluated).  This commit introduces that conditional statement, and
also modifies all references to $stomp_pool to reflect
$stomp_pool_real (which holds the evaluated, and defaulted-if-necessary,
value of $stomp_pool).