Parameterize MCollective's confdir
Previously, the path to MCollective's configuration directory was
being hard-coded in the params class. This commit introduces a new
parameter, mc_confdir, which points to MCollective's configuration
directory.  This commit also modifies any parameters that were
currently using hard-coded configuration directory values and changes
them to use the new mc_confdir parameter.

The problem, again, is that parameters in the parameter declaration
section of a class cannot rely on the values of other parameters due
to a bug in parse order. Because of this, any parameter that needs
to use the new mc_confdir parameter must be defaulted to 'UNSET' and
checked in the body of the class.