Commit the mcollective::user defined type
This commit will introduce a new defined resource type called
mcollective::user that can be used to create a local mcollective
user.  This user will contain the certificates necessary to send
messages to the MCollective collective.  Currently, this type assumes
that you are either a CA or a Puppet Master.